Preview version

+++ data/src/main/java/io/bsu/mmf/helpme/data/weather/CurrentWeather.kt
@@ -3,6 +3,6 @@ package io.bsu.mmf.helpme.data.weather
 data class CurrentWeather(
         val temp: Double,
         val feelTemp: Double,
-        val windSpeed: Double,
+        val windSpeed: Int,
         val weatherIcon: String
 )
